swagger refactorizare

+++ src/main/java/ro/pickupservice/controllers/location/bean/response/LocationResponse.java
@@ -6,21 +6,24 @@ public class LocationResponse implements Serializable{
 
 	private static final long serialVersionUID = 2481576516116919736L;
 	
-	private Long latitude;
-	private Long longitude;
+	private Float latitude;
+	private Float longitude;
 	
-	public Long getLatitude() {
+	public Float getLatitude() {
 		return latitude;
 	}
-	public void setLatitude(Long latitude) {
+	public void setLatitude(Float latitude) {
 		this.latitude = latitude;
 	}
-	public Long getLongitude() {
+
+	public Float getLongitude() {
 		return longitude;
 	}
-	public void setLongitude(Long longitude) {
+
+	public void setLongitude(Float longitude) {
 		this.longitude = longitude;
 	}
+
 	@Override
 	public String toString() {
 		return "LocationResponse [latitude=" + latitude + ", longitude=" + longitude + "]";
